split query param and sqli moduel
