+++ examples/wasm/BUILD
@@ -15,11 +15,29 @@ cc_library(
     deps = ["@github_nlohmann_json//:json"],
 )
 
+cc_library(
+    name = "query_param",
+    srcs = ["query_param.cc"],
+    hdrs = ["query_param.h"],
+)
+
+cc_library(
+    name = "sqli",
+    srcs = ["sqli.cc"],
+    hdrs = ["sqli.h"],
+    deps = [
+        "@libinjection//:libinjection",
+        ":query_param",
+    ]
+)
+
 cc_test(
     name = "unit_test",
     srcs = ["unit_test.cc"],
     deps = [
         ":config",
+        ":query_param",
+        ":sqli",
         "@gtest//:gtest",
         "@gtest//:gtest_main"
     ],
@@ -29,9 +47,9 @@ wasm_cc_binary(
     srcs = ["envoy_filter_http_wasm_example.cc"],
     deps = [
         "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics",
-        "@libinjection//:libinjection",
         ":config",
-        # "@github_nlohmann_json//:json",
+        ":sqli",
+        ":query_param",
     ],
     linkstatic = 1,
 )
