+++ examples/wasm/unit_test.cc
@@ -1,4 +1,5 @@
 #include "gtest/gtest.h"
+
 #include "config.h"
 
 TEST(ConfigTest, DefaultConfig) {
